Planner_test passing:   - Testing if graph is isomorphic   - Testing action plan Old Planner.py delete Planner2.py renamed to Planner.py
